pyproject.toml
[project]
name = "platform"
version = "0.0.1"
dependencies = ["pygame-ce==2.5.3"]

[tool.setuptools]
packages = []